requirements.txt
streamlit>=1.49
Pillow>=11
